Rename RegularExpressions to Regexes throughout repository
Complete refactoring of the repository to rename Platform.RegularExpressions.Transformer
to Platform.Regexes.Transformer as requested in issue #37.

Changes include:
- Renamed all C# namespaces from Platform.RegularExpressions.Transformer to Platform.Regexes.Transformer
- Renamed project directories: Platform.RegularExpressions.Transformer → Platform.Regexes.Transformer
- Renamed project files: Platform.RegularExpressions.Transformer.csproj → Platform.Regexes.Transformer.csproj
- Updated assembly names, package IDs, and NuGet metadata in project files
- Updated solution file references to point to renamed projects
- Updated Python setup.py URLs to match new repository structure
- Updated README.md with all new URLs, package names, and documentation links
- Preserved System.Text.RegularExpressions imports (unchanged as they reference .NET framework)

Tests: All 9 tests pass, solution builds successfully with no errors

Fixes #37
