+++ README.md
@@ -1,15 +1,15 @@
-[![Actions Status](https://github.com/linksplatform/RegularExpressions.Transformer/workflows/CD/badge.svg)](https://github.com/linksplatform/RegularExpressions.Transformer/actions?workflow=CD)
-[![Codacy Badge](https://api.codacy.com/project/badge/Grade/7bcd272efb834b7993f0cf3ea1e9bb69)](https://www.codacy.com/manual/drakonard/RegularExpressions.Transformer?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=linksplatform/RegularExpressions.Transformer&amp;utm_campaign=Badge_Grade)
+[![Actions Status](https://github.com/linksplatform/Regexes.Transformer/workflows/CD/badge.svg)](https://github.com/linksplatform/Regexes.Transformer/actions?workflow=CD)
+[![Codacy Badge](https://api.codacy.com/project/badge/Grade/7bcd272efb834b7993f0cf3ea1e9bb69)](https://www.codacy.com/manual/drakonard/Regexes.Transformer?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=linksplatform/Regexes.Transformer&amp;utm_campaign=Badge_Grade)
 [![CodeFactor](https://www.codefactor.io/repository/github/linksplatform/regularexpressions.transformer/badge)](https://www.codefactor.io/repository/github/linksplatform/regularexpressions.transformer)
 
-| [![NuGet Version and Downloads count](https://img.shields.io/nuget/v/Platform.RegularExpressions.Transformer?label=nuget&style=flat)](https://www.nuget.org/packages/Platform.RegularExpressions.Transformer) | C# |
+| [![NuGet Version and Downloads count](https://img.shields.io/nuget/v/Platform.Regexes.Transformer?label=nuget&style=flat)](https://www.nuget.org/packages/Platform.Regexes.Transformer) | C# |
 |-|-|
 | [![PyPI version](https://badge.fury.io/py/retranslator.svg)](https://badge.fury.io/py/retranslator)  | __Python__  |
 | [![nimble](https://raw.githubusercontent.com/yglukhov/nimble-tag/master/nimble.png)](https://nimble.directory/pkg/retranslator)  | __Nim__  |
 
-# [RegularExpressions.Transformer](https://github.com/linksplatform/RegularExpressions.Transformer)
+# [Regexes.Transformer](https://github.com/linksplatform/Regexes.Transformer)
 
-LinksPlatform's Platform.RegularExpressions.Transformer Class Library.
+LinksPlatform's Platform.Regexes.Transformer Class Library.
 
 Another way to transform text is to use chain of regular expression in the `sed` command.
 
@@ -17,16 +17,16 @@ Another way to transform text is to use chain of regular expression in the `sed`
 sed -e 's/first match/first substution/g' -e 's/second match/second substution/g'
 ```
 
-Namespace: [Platform.RegularExpressions.Transformer](https://linksplatform.github.io/RegularExpressions.Transformer/csharp/api/Platform.RegularExpressions.Transformer.html)
+Namespace: [Platform.Regexes.Transformer](https://linksplatform.github.io/Regexes.Transformer/csharp/api/Platform.Regexes.Transformer.html)
 
 Forked from: [LinksPlatform/Collections.Methods/CSharpToCppTranslator](https://github.com/linksplatform/Collections.Methods/tree/20ac8963cdeb8f68013139f4083abd98be03ff43/CSharpToCppTranslator)
 
-NuGet package: [Platform.RegularExpressions.Transformer](https://www.nuget.org/packages/Platform.RegularExpressions.Transformer)
+NuGet package: [Platform.Regexes.Transformer](https://www.nuget.org/packages/Platform.Regexes.Transformer)
 
-Python version: [retranslator](https://github.com/linksplatform/RegularExpressions.Transformer/tree/master/python)
+Python version: [retranslator](https://github.com/linksplatform/Regexes.Transformer/tree/master/python)
 
-## [Documentation](https://linksplatform.github.io/RegularExpressions.Transformer)
-[PDF file](https://linksplatform.github.io/RegularExpressions.Transformer/csharp/Platform.RegularExpressions.Transformer.pdf) with code for e-readers.
+## [Documentation](https://linksplatform.github.io/Regexes.Transformer)
+[PDF file](https://linksplatform.github.io/Regexes.Transformer/csharp/Platform.Regexes.Transformer.pdf) with code for e-readers.
 
 ## Dependent libraries
-*   [Platform.RegularExpressions.Transformer.CSharpToCpp](https://github.com/linksplatform/RegularExpressions.Transformer.CSharpToCpp)
+*   [Platform.Regexes.Transformer.CSharpToCpp](https://github.com/linksplatform/Regexes.Transformer.CSharpToCpp)
