+++ csharp/Platform.Regexes.Transformer.Tests/SubstitutionRuleTests.cs
@@ -1,7 +1,7 @@
 using System.Text.RegularExpressions;
 using Xunit;
 
-namespace Platform.RegularExpressions.Transformer.Tests
+namespace Platform.Regexes.Transformer.Tests
 {
     /// <summary>
     /// <para>
