+++ csharp/Platform.Regexes.Transformer/FileTransformer.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 
 #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 
-namespace Platform.RegularExpressions.Transformer
+namespace Platform.Regexes.Transformer
 {
     /// <summary>
     /// <para>
