+++ csharp/Platform.Regexes.Transformer/IFileTransformer.cs
@@ -1,25 +1,25 @@
 using System.Runtime.CompilerServices;
-using System.Text.RegularExpressions;
 
 #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 
-namespace Platform.RegularExpressions.Transformer
+namespace Platform.Regexes.Transformer
 {
     /// <summary>
     /// <para>
-    /// Defines the substitution rule.
+    /// Defines the file transformer.
     /// </para>
     /// <para></para>
     /// </summary>
-    public interface ISubstitutionRule
+    /// <seealso cref="ITransformer"/>
+    public interface IFileTransformer : ITransformer
     {
         /// <summary>
         /// <para>
-        /// Gets the match pattern value.
+        /// Gets the source file extension value.
         /// </para>
         /// <para></para>
         /// </summary>
-        Regex MatchPattern
+        string SourceFileExtension
         {
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
             get;
@@ -27,11 +27,11 @@ namespace Platform.RegularExpressions.Transformer
 
         /// <summary>
         /// <para>
-        /// Gets the substitution pattern value.
+        /// Gets the target file extension value.
         /// </para>
         /// <para></para>
         /// </summary>
-        string SubstitutionPattern
+        string TargetFileExtension
         {
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
             get;
@@ -39,14 +39,19 @@ namespace Platform.RegularExpressions.Transformer
 
         /// <summary>
         /// <para>
-        /// Gets the maximum repeat count value.
+        /// Transforms the source path.
         /// </para>
         /// <para></para>
         /// </summary>
-        int MaximumRepeatCount
-        {
-            [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            get;
-        }
+        /// <param name="sourcePath">
+        /// <para>The source path.</para>
+        /// <para></para>
+        /// </param>
+        /// <param name="targetPath">
+        /// <para>The target path.</para>
+        /// <para></para>
+        /// </param>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        void Transform(string sourcePath, string targetPath);
     }
 }
