+++ csharp/Platform.Regexes.Transformer/ITextTransformersListExtensions.cs
@@ -5,7 +5,7 @@ using Platform.Collections;
 
 #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 
-namespace Platform.RegularExpressions.Transformer
+namespace Platform.Regexes.Transformer
 {
     /// <summary>
     /// <para>
