+++ csharp/Platform.Regexes.Transformer/Platform.Regexes.Transformer.csproj
@@ -1,20 +1,20 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <Description>LinksPlatform's Platform.RegularExpressions.Transformer Class Library</Description>
+    <Description>LinksPlatform's Platform.Regexes.Transformer Class Library</Description>
     <Copyright>Konstantin Diachenko</Copyright>
-    <AssemblyTitle>Platform.RegularExpressions.Transformer</AssemblyTitle>
+    <AssemblyTitle>Platform.Regexes.Transformer</AssemblyTitle>
     <VersionPrefix>0.3.4</VersionPrefix>
     <Authors>Konstantin Diachenko</Authors>
     <TargetFramework>net8</TargetFramework>
-    <AssemblyName>Platform.RegularExpressions.Transformer</AssemblyName>
-    <PackageId>Platform.RegularExpressions.Transformer</PackageId>
-    <PackageTags>LinksPlatform;RegularExpressions.Transformer;IContext;ISubstitutionRule;ITransformer;Context;RegexExtensions;SubstitutionRule;Transformer;TransformerCLI</PackageTags>
+    <AssemblyName>Platform.Regexes.Transformer</AssemblyName>
+    <PackageId>Platform.Regexes.Transformer</PackageId>
+    <PackageTags>LinksPlatform;Regexes.Transformer;IContext;ISubstitutionRule;ITransformer;Context;RegexExtensions;SubstitutionRule;Transformer;TransformerCLI</PackageTags>
     <PackageIconUrl>https://raw.githubusercontent.com/linksplatform/Documentation/18469f4d033ee9a5b7b84caab9c585acab2ac519/doc/Avatar-rainbow-icon-64x64.png</PackageIconUrl>
-    <PackageProjectUrl>https://linksplatform.github.io/RegularExpressions.Transformer</PackageProjectUrl>
+    <PackageProjectUrl>https://linksplatform.github.io/Regexes.Transformer</PackageProjectUrl>
     <PackageLicenseExpression>Unlicensed</PackageLicenseExpression>
     <RepositoryType>git</RepositoryType>
-    <RepositoryUrl>git://github.com/linksplatform/RegularExpressions.Transformer</RepositoryUrl>
+    <RepositoryUrl>git://github.com/linksplatform/Regexes.Transformer</RepositoryUrl>
     <GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
     <GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
