+++ csharp/Platform.Regexes.Transformer/Steps.cs
@@ -3,7 +3,7 @@ using System.Runtime.CompilerServices;
 
 #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 
-namespace Platform.RegularExpressions.Transformer
+namespace Platform.Regexes.Transformer
 {
     /// <summary>
     /// <para>
