+++ csharp/Platform.Regexes.Transformer/StringExtensions.cs
@@ -4,7 +4,7 @@ using System.Text;
 
 #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 
-namespace Platform.RegularExpressions.Transformer
+namespace Platform.Regexes.Transformer
 {
     /// <summary>
     /// <para>
