+++ csharp/Platform.Regexes.Transformer/SubstitutionRule.cs
@@ -5,7 +5,7 @@ using System.Text.RegularExpressions;
 
 #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 
-namespace Platform.RegularExpressions.Transformer
+namespace Platform.Regexes.Transformer
 {
     /// <summary>
     /// <para>
